refixing adapter and sync with main
