+++ app/src/main/java/com/example/lokerin/ListJobAdapter.java
@@ -22,17 +22,15 @@ import com.google.firebase.database.ValueEventListener;
 import java.util.ArrayList;
 import java.util.List;
 
-public class ListJobAdapter extends RecyclerView.Adapter<ListJobAdapter.CardViewHolder> implements Filterable {
+public class ListJobAdapter extends RecyclerView.Adapter<ListJobAdapter.CardViewHolder>{
 
     private List<JobData> jobDataList;
-    private List<JobData> filteredList;
     private Context context;
     private FirebaseDatabase firebaseDatabase;
 
     public ListJobAdapter(Context context, List<JobData> jobDataList) {
         this.context = context;
         this.jobDataList = jobDataList;
-        this.filteredList = new ArrayList<>(jobDataList);
     }
 
     @NonNull
@@ -44,11 +42,11 @@ public class ListJobAdapter extends RecyclerView.Adapter<ListJobAdapter.CardView
 
     @Override
     public void onBindViewHolder(@NonNull CardViewHolder holder, int position) {
-        JobData data = filteredList.get(position); // Use filteredList
-        holder.jobTitle.setText(data.getJobTitle());
-        holder.jobLocation.setText(data.getJobLocation());
-        holder.jobDateUpload.setText(data.getJobDateUpload());
-        holder.jobCategory.setText(data.getJobCategory());
+        JobData data = jobDataList.get(position);
+        holder.jobTitle.setText(data.getJobTitle() != null ? data.getJobTitle() : "-");
+        holder.jobLocation.setText(data.getJobProvince() != null ? data.getJobProvince() : "-");
+        holder.jobDateUpload.setText(data.getJobDateUpload() != null ? data.getJobDateUpload() : "-");
+        holder.jobCategory.setText(data.getJobCategory() != null ? data.getJobCategory() : "-");
 
         holder.itemView.setOnClickListener(v -> {
             FirebaseAuth mAuth = FirebaseAuth.getInstance();
@@ -59,23 +57,13 @@ public class ListJobAdapter extends RecyclerView.Adapter<ListJobAdapter.CardView
                 public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
                     String userType = dataSnapshot.child("type").getValue().toString();
                     if(userType.equals("pelanggan")) {
-                        Intent intent = new Intent(context, PelangganDetailJobActivity.class);
-                        intent.putExtra("jobTitle", data.getJobTitle());
-                        intent.putExtra("jobLocation", data.getJobLocation());
-                        intent.putExtra("jobDateUpload", data.getJobDateUpload());
-                        intent.putExtra("jobCategory", data.getJobCategory());
-                        intent.putExtra("jobStatus", data.getJobStatus());
-                        intent.putExtra("jobApplicants", data.getApplicants().toArray(new User[0]));
+                        Intent intent = new Intent(context.getApplicationContext(), PelangganDetailJobActivity.class);
+                        intent.putExtra("jobId", data.getJobId());
                         context.startActivity(intent);
                     }
                     else {
-                        Intent intent = new Intent(context, PekerjaDetailJobActivity.class);
-                        intent.putExtra("jobTitle", data.getJobTitle());
-                        intent.putExtra("jobLocation", data.getJobLocation());
-                        intent.putExtra("jobDateUpload", data.getJobDateUpload());
-                        intent.putExtra("jobCategory", data.getJobCategory());
-                        intent.putExtra("jobStatus", data.getJobStatus());
-                        intent.putExtra("jobApplicants", data.getApplicants().toArray(new User[0]));
+                        Intent intent = new Intent(context.getApplicationContext(), PekerjaDetailJobActivity.class);
+                        intent.putExtra("jobId", data.getJobId());
                         context.startActivity(intent);
                     }
                 }
@@ -85,51 +73,17 @@ public class ListJobAdapter extends RecyclerView.Adapter<ListJobAdapter.CardView
 
                 }
             });
-//            Intent intent = new Intent(context, PekerjaDetailJobActivity.class);
-//            intent.putExtra("jobTitle", data.getJobTitle());
-//            intent.putExtra("jobLocation", data.getJobLocation());
-//            intent.putExtra("jobDateUpload", data.getJobDateUpload());
-//            intent.putExtra("jobCategory", data.getJobCategory());
-//            intent.putExtra("jobStatus", data.getJobStatus());
-//            intent.putExtra("jobApplicants", data.getApplicants().toArray(new User[0]));
-//            context.startActivity(intent);
         });
     }
 
     @Override
     public int getItemCount() {
-        return filteredList.size();
+        return jobDataList.size();
     }
 
-    @Override
-    public Filter getFilter() {
-        return new Filter() {
-            @Override
-            protected FilterResults performFiltering(CharSequence constraint) {
-                String query = constraint.toString().toLowerCase().trim();
-                List<JobData> filtered = new ArrayList<>();
-
-                if (query.isEmpty()) {
-                    filtered.addAll(jobDataList);
-                } else {
-                    for (JobData job : jobDataList) {
-                        if (job.getJobTitle().toLowerCase().contains(query)) {
-                            filtered.add(job);
-                        }
-                    }
-                }
-                FilterResults results = new FilterResults();
-                results.values = filtered;
-                return results;
-            }
-
-            @Override
-            protected void publishResults(CharSequence constraint, FilterResults results) {
-                filteredList.clear();
-                filteredList.addAll((List<JobData>) results.values);
-                notifyDataSetChanged();
-            }
-        };
+    public void updateList(List<JobData> updatedList) {
+        this.jobDataList = updatedList;
+        notifyDataSetChanged();
     }
 
     public static class CardViewHolder extends RecyclerView.ViewHolder {
