+++ app/src/main/java/com/example/lokerin/PekerjaRecommendJobActivity.java
@@ -42,7 +42,7 @@ public class PekerjaRecommendJobActivity extends AppCompatActivity {
     private String currentUserId;
 
     private List<JobData> jobDataList = new ArrayList<>();
-    private ListJobPekerjaAdapter adapter;
+    private ListJobAdapter adapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -74,7 +74,7 @@ public class PekerjaRecommendJobActivity extends AppCompatActivity {
 
         rvJobs = findViewById(R.id.recyclerView);
         rvJobs.setLayoutManager(new LinearLayoutManager(this));
-        adapter = new ListJobPekerjaAdapter(this, jobDataList);
+        adapter = new ListJobAdapter(this, jobDataList);
         rvJobs.setAdapter(adapter);
 
         etSearchBar = findViewById(R.id.search_bar);
